Move enum out of tests directory
