+++ tests/Browser/SupportEnums/Test.php
@@ -13,7 +13,7 @@ class Test extends TestCase
     public function test()
     {
         $this->browse(function ($browser) {
-            Livewire::visit($browser, 'Tests\Browser\SupportEnums\Component')
+            Livewire::visit($browser, Component::class)
                 ->assertSee('Be excellent to each other');
         });
     }
